rename files to commonjs
