+++ src/svelteconfig.cjs
